backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
PyJWT
transformers
python-multipart
